Renamed swagger/__init__.py and deleted swagger folder
