+++ vantage6-server/vantage6/server/resource/swagger_templates.py
